djangoProject/requirements.txt
Django~=5.1.2
pytz~=2024.2
jwt~=1.3.1
requests~=2.32.3
djangorestframework~=3.15.2